pyproject.toml
[tool.poetry]
name = "bingo-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "bingo_api"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.98.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
